Split strata.rs into api/ directory module for scalable executor structure
Renames the module from `strata` to `api` to reflect the executor's role
as the standard command layer for Python/Node/CLI/MCP. Splits the 1,626-line
monolith into 8 focused files (mod.rs + 7 domain files) while preserving
the identical public API surface.
