+++ crates/executor/src/api/db.rs
@@ -1,0 +1,51 @@
+//! Database operations: ping, info, flush, compact.
+
+use super::Strata;
+use crate::{Command, Error, Output, Result};
+use crate::types::*;
+
+impl Strata {
+    // =========================================================================
+    // Database Operations (4)
+    // =========================================================================
+
+    /// Ping the database.
+    pub fn ping(&self) -> Result<String> {
+        match self.executor.execute(Command::Ping)? {
+            Output::Pong { version } => Ok(version),
+            _ => Err(Error::Internal {
+                reason: "Unexpected output for Ping".into(),
+            }),
+        }
+    }
+
+    /// Get database info.
+    pub fn info(&self) -> Result<DatabaseInfo> {
+        match self.executor.execute(Command::Info)? {
+            Output::DatabaseInfo(info) => Ok(info),
+            _ => Err(Error::Internal {
+                reason: "Unexpected output for Info".into(),
+            }),
+        }
+    }
+
+    /// Flush the database to disk.
+    pub fn flush(&self) -> Result<()> {
+        match self.executor.execute(Command::Flush)? {
+            Output::Unit => Ok(()),
+            _ => Err(Error::Internal {
+                reason: "Unexpected output for Flush".into(),
+            }),
+        }
+    }
+
+    /// Compact the database.
+    pub fn compact(&self) -> Result<()> {
+        match self.executor.execute(Command::Compact)? {
+            Output::Unit => Ok(()),
+            _ => Err(Error::Internal {
+                reason: "Unexpected output for Compact".into(),
+            }),
+        }
+    }
+}
